Moving the functions to tests/common/cisco_data.py.
